+++ src/Ch03/NestedLets.hs
@@ -1,3 +1,4 @@
+{-# OPTIONS_GHC -Wno-name-shadowing -Wno-unused-matches #-}
 -- |
 -- Module      :  Ch03.NestedLets
 -- Description :  Shadowing
@@ -12,16 +13,19 @@
 
 module Ch03.NestedLets where
 
--- Shadowing, p. 62
+-- aded GHC pragma to remove warns -Wname-shadowing -Wunused-matches
 
+-- Shadowing, p. 62
+foo :: Integer                  -- added to remove [-Wmissing-signatures] and [-Wtype-defaults]
 foo = let a = 1
           in let b = 2
               in a + b
 
+bar :: ([Char], Integer)        -- added to remove [-Wmissing-signatures] and [-Wtype-defaults]
 bar = let x = 1
-          in ((let x = "foo" in x), x)
+          in ((let x = "foo" in x), x)  -- Complaints about [-Wname-shadowing]
 
 -- this function can be any type
--- @quux :: t -> [Char]
-quux a = let a = "foo"
+quux :: t -> [Char]                     -- Complaints [-Wname-shadowing] and [-Wunused-matches]
+quux a = let a = "foo"                  
           in a ++ "eek!"